code/requirements.txt
aenum==3.1.12
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dnspython==2.3.0
exceptiongroup==1.1.1
fastapi==0.97.0
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
h11==0.14.0
hydra-core==1.3.2
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lightly==1.4.8
lightly-utils==0.0.2
lightning-utilities==0.8.0
MarkupSafe==2.1.3
matplotlib==3.7.1
motor==3.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.3
odmantic==0.9.2
omegaconf==2.3.0
opencv-python==4.7.0.72
packaging==23.1
path==16.6.0
Pillow==9.5.0
pydantic==1.10.9
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==2.0.3
PyYAML==6.0
requests==2.31.0
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
threadpoolctl==3.1.0
torch==2.0.1
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
yarl==1.9.2
zipp==3.15.0
